iiot-monitoring-simmk/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
pydantic>=2.6,<3.0
httpx>=0.27,<1.0
streamlit>=1.32,<2.0
pandas>=2.2,<3.0
requests>=2.31,<3.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
aiosqlite>=0.20,<1.0
paho-mqtt>=2.1,<3.0
prometheus-client>=0.20,<1.0
python-json-logger>=2.0,<4.0
pytest>=8.2,<9.0
pytest-asyncio>=0.23,<1.0
ruff>=0.5,<1.0
